fixed bug that made options dissapear
Was reassigning the list of options found to the same variable, which would hold it across to the next zcta.  So once one set was removed because no data, it never came back since the code assigning the initial values is global in the file and not in the function.  Change it to store the initial dictionary in one variable and assign to another when removing items.